feat: interactive live api execution
